@@ -10,28 +10,28 @@
     "screenshots": "concurrently \"vite dev\" \"node scripts/generate-screenshots.js\""
   },
   "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "^3.1.2",
+    "@sveltejs/vite-plugin-svelte": "^5.0.3",
     "@tsconfig/svelte": "^5.0.4",
     "@types/cors": "^2.8.17",
     "@types/express": "^4.17.21",
     "@types/react": "^18.3.12",
     "@types/react-dom": "^18.3.1",
-    "@vitejs/plugin-react": "^4.3.3",
-    "concurrently": "^9.0.1",
+    "@vitejs/plugin-react": "^4.3.4",
+    "concurrently": "^9.1.2",
     "cors": "^2.8.5",
-    "express": "^4.21.1",
+    "express": "^4.21.2",
     "glob": "^11.0.0",
     "picocolors": "^1.1.1",
-    "svelte-check": "^4.0.5",
-    "tslib": "^2.8.0",
-    "typescript": "^5.5.4",
-    "vite": "^5.4.11"
+    "svelte-check": "^4.1.3",
+    "tslib": "^2.8.1",
+    "typescript": "^5.7.3",
+    "vite": "^6.0.7"
   },
   "dependencies": {
     "@dagrejs/dagre": "^1.1.4",
     "@threlte/core": "^7.3.1",
-    "@xyflow/react": "^12.3.2",
-    "@xyflow/svelte": "^0.1.21",
+    "@xyflow/react": "^12.3.6",
+    "@xyflow/svelte": "file:.yalc/@xyflow/svelte",
     "d3-drag": "^3.0.0",
     "d3-force": "^3.0.0",
     "d3-hierarchy": "^3.1.2",
@@ -40,17 +40,17 @@
     "elkjs": "^0.9.3",
     "entitree-flex": "^0.4.1",
     "html-to-image": "^1.11.11",
-    "nanoid": "^5.0.7",
+    "nanoid": "^5.0.9",
     "puppeteer": "^22.7.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-icons": "^5.3.0",
+    "react-icons": "^5.4.0",
     "react-remark": "^2.1.0",
-    "styled-components": "^6.1.13",
-    "svelte": "^4.2.19",
-    "svelte-feather-icons": "^4.1.0",
+    "styled-components": "^6.1.14",
+    "svelte": "^5.17.2",
+    "svelte-feather-icons": "^4.2.0",
     "twind": "^0.16.19",
     "webcola": "^3.4.0",
-    "zustand": "5.0.0"
+    "zustand": "5.0.3"
   }
 }
