@@ -3,28 +3,29 @@
 
   const connection = useConnection();
 
-  let path = '';
-
-  $: {
-    const { from, to } = $connection;
-    path = `M${from.x},${from.y} C ${from.x} ${to.y} ${from.x} ${to.y} ${to.x},${to.y}`;
-  }
+  let path: string | null = $derived.by(() => {
+    if (connection.current.inProgress) {
+      const { from, to } = connection.current;
+      return `M${from.x},${from.y} C ${from.x} ${to.y} ${from.x} ${to.y} ${to.x},${to.y}`;
+    }
+    return null;
+  });
 </script>
 
-{#if path}
+{#if connection.current.inProgress}
   <path
     fill="none"
     stroke-width={1.5}
     class="animated"
-    stroke={$connection.fromHandle?.id}
+    stroke={connection.current.fromHandle.id}
     d={path}
   />
   <circle
-    cx={$connection.to.x}
-    cy={$connection.to.y}
+    cx={connection.current.to.x}
+    cy={connection.current.to.y}
     fill="#fff"
     r={3}
-    stroke={$connection.fromHandle?.id}
+    stroke={connection.current.fromHandle.id}
     stroke-width={1.5}
   />
 {/if}
