@@ -1,6 +1,11 @@
 <script lang="ts">
-  import { writable } from 'svelte/store';
-  import { SvelteFlow, Background, BackgroundVariant, type Edge, type Node } from '@xyflow/svelte';
+  import {
+    SvelteFlow,
+    Background,
+    BackgroundVariant,
+    type Edge,
+    type Node,
+  } from '@xyflow/svelte';
 
   import CustomNode from './CustomNode.svelte';
   import ConnectionLine from './ConnectionLine.svelte';
@@ -8,24 +13,26 @@
   import '@xyflow/svelte/dist/style.css';
 
   const nodeTypes = {
-    custom: CustomNode
+    custom: CustomNode,
   };
 
-  const nodes = writable<Node[]>([
+  let nodes = $state.raw<Node[]>([
     {
       id: 'connectionline-1',
       type: 'custom',
       data: { label: 'Node 1' },
-      position: { x: 250, y: 5 }
-    }
+      position: { x: 250, y: 5 },
+    },
   ]);
 
-  const edges = writable<Edge[]>([]);
+  let edges = $state.raw<Edge[]>([]);
 </script>
 
 <div style="height:100vh;">
-  <SvelteFlow {nodeTypes} {nodes} {edges} fitView>
-    <ConnectionLine slot="connectionLine" />
+  <SvelteFlow bind:nodes bind:edges {nodeTypes} fitView>
+    {#snippet connectionLine()}
+      <ConnectionLine />
+    {/snippet}
     <Background variant={BackgroundVariant.Lines} />
   </SvelteFlow>
 </div>
