@@ -1,7 +1,8 @@
+import { mount } from 'svelte';
 import App from './App.svelte';
 
 import './styles.css';
 
-const app = new App({
-  target: document.querySelector('#app'),
+mount(App, {
+  target: document.getElementById('app')!,
 });
